requirements.txt
fastapi>=0.95.0
confluent-kafka>=2.1.0
weaviate-client>=3.21.0,<4.0.0
boto3>=1.26.0
pydantic>=1.10.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.8.4
requests>=2.28.0
typing-extensions>=4.5.0
uvicorn[standard]>=0.22.0
